chore(whoop): mypy + ruff cleanup on new WHOOP modules
- Type fixes so all 14 new core modules are mypy-clean (max_hr band float, behaviour
  float cast, strain_target state narrowing, cosinor/dfa float lists)
- ruff I001 import-sort fix in test_training_load

Note: repo-wide Code Quality debt (main.py I001, ~20 pre-existing black files, pre-existing
mypy in api_key_repo + _longest_low_block) predates this branch and remains — see PR discussion.

+++ rivaflow/rivaflow/core/behaviour.py
@@ -20,7 +20,7 @@ def cohens_d(a: list[float], b: list[float]) -> float | None:
     pooled = (((na - 1) * va + (nb - 1) * vb) / (na + nb - 2)) ** 0.5
     if pooled == 0:
         return None
-    return (mean(a) - mean(b)) / pooled
+    return float((mean(a) - mean(b)) / pooled)
 
 
 def _magnitude(d: float) -> str:
